requirements.txt
Django>=5.0,<6.0
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
anthropic>=0.40
requests>=2.31
playwright>=1.40

gunicorn>=22.0
whitenoise>=6.6
python-dotenv>=1.0
psycopg2-binary>=2.9
